+++ modules/search-results/inc/hit-flore.php
@@ -1,7 +1,7 @@
 <?php function telabotanica_module_search_results_hit_flore($hit) {
 	printf(
 		'<a class="search-results-hit-link" href="%s" title="%s">',
-		$hit['bdtfx']['url'],
+		$hit['bdtfx']['permalink'],
 		$hit['bdtfx']['scientific_name']
 	);
 		printf(