+++ modules/share/script.js
@@ -1,10 +1,9 @@
-'use strict';
-
 var Tela = window.Tela || {};
+Tela.modules = Tela.modules || {};
 
-Tela.share = (function(){
+Tela.modules.share = (function(){
 
-  function share(selector){
+  function module(selector){
     var $el     = $(selector),
       $links;
 
@@ -32,12 +31,12 @@ Tela.share = (function(){
 
   return function(selector){
     return $(selector).each(function(){
-      share(this);
+      module(this);
     });
   };
 
 })();
 
 $(document).ready(function(){
-  Tela.share('.share');
+  Tela.modules.share('.share');
 });