+++ search.php
@@ -1,6 +1,29 @@
 <?php
+
+$search_query = sanitize_text_field( get_search_query() );
+
 // Are we searching in a specific index?
-$current_index = get_query_var( 'index', false );
+$current_index = sanitize_key( get_query_var( 'in', false ) );
+
+// Redirect some indices to the category or page, while preserving the query
+switch ($current_index) {
+	case 'actualites':
+		$redirect_url = get_category_link( get_category_by_slug( 'actualites' ) );
+		break;
+
+	case 'evenements':
+		$redirect_url = get_category_link( get_category_by_slug( 'evenements' ) );
+		break;
+
+	case 'projets':
+		$redirect_url = get_permalink(get_page_by_path( 'projets' ));
+		break;
+}
+
+if ( isset($redirect_url) ) {
+	wp_redirect( $redirect_url . '?q=' . $search_query );
+	exit;
+}
 
 // Force a small header (without use cases navigation)
 $header_small = true;
@@ -12,28 +35,38 @@ get_header();
 		<main id="main" class="site-main">
 
 			<?php
+			if ( telabotanica_algolia_check(true) ) :
 
-			// Get Algolia instance
-			if ( class_exists( 'Algolia_Plugin' ) ) :
-				$algolia = Algolia_Plugin::get_instance();
+				// Get Algolia instance
+				$algolia_client = telabotanica_algolia_client();
+				$algolia_autocomplete_config = telabotanica_algolia_config()['autocomplete'];
 
 				if ( $current_index ) :
 
 					// Retrieve the label for the current index
-					$indices = $algolia->get_autocomplete_config()->get_config();
+					$indices = $algolia_autocomplete_config['sources'];
 					foreach ( $indices as $index ) :
 						if ( $index['index_id'] === $current_index ) {
-							$current_index_label = $index['label'];
-							$current_index_name = $index['index_name'];
+							$current_index = [
+								'id' => $current_index,
+								'label' => $index['label'],
+								'name' => $index['index_name'],
+								'filters' => @$index['filters'] ?: []
+							];
+							break;
 						}
 					endforeach;
 
 					// Perform the search
-					$index = $algolia->get_api()->get_client()->initIndex($current_index_name);
-					$results = $index->search(get_search_query());
+					$index = $algolia_client->initIndex($current_index['name']);
+					$results = $index->search($search_query, [
+						'hitsPerPage' => 20
+					]);
 
 					the_telabotanica_module('cover-search', [
-						'total_results' => false
+						'index' => $current_index['id'],
+						'total_results' => false,
+						'instantsearch' => true
 					]);
 					?>
 
@@ -41,7 +74,9 @@ get_header();
 						<div class="layout-wrapper">
 								<aside class="layout-column">
 									<?php
-									the_telabotanica_module('search-filters');
+									the_telabotanica_module('search-filters', [
+										'filters' => $current_index['filters']
+									]);
 									the_telabotanica_module('button-top');
 									?>
 								</aside>
@@ -50,43 +85,48 @@ get_header();
 									the_telabotanica_module('breadcrumbs', [
 										'items' => [
 											[
-												'href' => esc_url( home_url( '/' ) . '?s=' . get_search_query() ),
+												'href' => esc_url( home_url( '/' ) . '?s=' . $search_query ),
 												'text' => __('Recherche', 'telabotanica')
 											],
 											[
-												'text' => $current_index_label
+												'text' => $current_index['label']
 											],
-											[ 'text' => sprintf( _n(
+											[ 'text' => '<span id="search-stats">' . sprintf( _n(
 												'%s résultat trouvé',
 												'%s résultats trouvés',
 												$results['nbHits'],
 												'telabotanica'
-											), number_format_i18n( $results['nbHits'] ) ) ]
+											), number_format_i18n( $results['nbHits'] ) ) . '</span>' ]
 										]
 									]);
 
-									var_dump($results);
+									echo '<div id="search-hits">';
+										foreach ($results['hits'] as $hit) {
+											$hit['type'] = $current_index['id'];
+											the_telabotanica_module('search-hit', $hit);
+										}
+									echo '</div>';
+
+									// TODO: add pagination
 									?>
 								</div>
 							</div>
 						</div>
-
 				<?php
 				else :
 					// Perform several queries in a single API call
 					$queries = [];
-					$indices = $algolia->get_autocomplete_config()->get_config();
 
-					foreach ( $indices as $index ) :
+					foreach ( $algolia_autocomplete_config['sources'] as $index ) :
 						$queries[] = [
 							'indexName' => $index['index_name'],
-							'query' => get_search_query(),
-							'hitsPerPage' => $index['max_suggestions'] * 2,
-							'facetFilters' => array_key_exists('default_facet_filters', $index) ? $index['default_facet_filters'] : null
+							'query' => $search_query,
+							'hitsPerPage' => $index['settings']['hitsPerPage'] * 2,
+							'facetFilters' => array_key_exists('facetFilters', $index['settings']) ? $index['settings']['facetFilters'] : null
 						];
 					endforeach;
 
-					$results = $algolia->get_api()->get_client()->multipleQueries($queries);
+					$results = $algolia_client->multipleQueries($queries);
 					$total_results = array_sum(array_column($results['results'], 'nbHits'));
 
 					the_telabotanica_module('cover-search', [